Added the ability to say "interface=True" in the Meta declaration.
"interface=True" means that this Model won't be instantiated, but it will be
derived from and the derived classes will override Meta.  Without this
style, errors were thrown.

This case could, in theory be avoided, but I think it would require a fair
amount of work.  For example, imagine a whole set of different document
classes that wanted to share a common override of the save() method.  You
might be able to get that correct with multiple inheritance, but I think
this way is actually more straightforward.